+++ test/client/functions/batch/hisDeleteByFilter.js
@@ -7,9 +7,9 @@
 
 const stubs = require('../../../stubs');
 const sinon = require('sinon');
+const {RequestError} = require("../../../../src/errors");
 const expect = require('chai').expect;
 const getInstance = stubs.getInstance;
-const Hs = require("../../../../src/utils/haystack");
 
 const HIS_DELETE_ENTITY_BATCH_SIZE = 100;
 const HIS_READ_LARGE_TIME_SERIES = {
@@ -36,9 +36,7 @@ const TIME_END = new Date();
 function genEntities(source, num) {
     const entities = [];
     for (let i = 0; i < num; i++) {
-        entities.push({
-            id: `r:${source[i]}`
-        });
+        entities.push(`id-${source[i]}`);
     }
 
     return entities;
@@ -50,7 +48,7 @@ describe("client.batch.hisDeleteByFilter", () => {
         http = new stubs.StubHTTPClient();
         log = new stubs.StubLogger();
         ws = getInstance(http, log);
-        ws.v2.find = sinon.stub().callsFake(() => []);
+        ws.findAsId = sinon.stub().callsFake(() => []);
         ws.hisRead = sinon.stub().callsFake((ids, start, end, batchSize) =>  {
             return {
                 rows: [
@@ -73,13 +71,13 @@ describe("client.batch.hisDeleteByFilter", () => {
             it("should create 1 request", async () => {
                 const entities = genEntities(
                     HIS_READ_LARGE_ENTITIES.ids, HIS_DELETE_ENTITY_BATCH_SIZE - 4);
-                ws.v2.find = sinon.stub().callsFake(() => entities);
+                ws.findAsId = sinon.stub().callsFake(() => [...entities]);
                 ws.batch.hisRead = sinon.stub().callsFake(() => HIS_READ_LARGE_ENTITIES.data);
                 await ws.batch.hisDeleteByFilter("point", TIME_START, TIME_END);
-                expect(ws.v2.find.calledOnce).to.be.true;
+                expect(ws.findAsId.calledOnce).to.be.true;
                 expect(ws.hisDelete.calledOnce).to.be.true;
                 expect(ws.hisDelete.args[0]).to.eql([
-                    entities.map((entity) => Hs.getId(entity)),
+                    entities,
                     "s:1970-01-01T00:00:00.000Z,1970-01-01T00:00:00.010Z"
                 ]);
             });
@@ -89,21 +87,17 @@ describe("client.batch.hisDeleteByFilter", () => {
             it("should create more than 1 request", async () => {
                 const entities = genEntities(
                     HIS_READ_LARGE_ENTITIES.ids, HIS_DELETE_ENTITY_BATCH_SIZE + 10);
-                ws.v2.find = sinon.stub().callsFake(() => entities);
+                ws.findAsId = sinon.stub().callsFake(() => [...entities]);
                 ws.batch.hisRead = sinon.stub().callsFake(() => HIS_READ_LARGE_ENTITIES.data);
                 await ws.batch.hisDeleteByFilter("point", TIME_START, TIME_END);
-                expect(ws.v2.find.calledOnce).to.be.true;
+                expect(ws.findAsId.calledOnce).to.be.true;
                 expect(ws.hisDelete.calledTwice).to.be.true;
                 expect(ws.hisDelete.args[0]).to.eql([
-                    entities
-                        .slice(0, HIS_DELETE_ENTITY_BATCH_SIZE)
-                        .map((entity) => Hs.getId(entity)),
+                    entities.slice(0, HIS_DELETE_ENTITY_BATCH_SIZE),
                     "s:1970-01-01T00:00:00.000Z,1970-01-01T00:00:00.010Z"
                 ]);
                 expect(ws.hisDelete.args[1]).to.eql([
-                    entities
-                        .slice(HIS_DELETE_ENTITY_BATCH_SIZE)
-                        .map((entity) => Hs.getId(entity)),
+                    entities.slice(HIS_DELETE_ENTITY_BATCH_SIZE),
                     "s:1970-01-01T00:00:00.000Z,1970-01-01T00:00:00.010Z"
                 ]);
             });
@@ -115,15 +109,15 @@ describe("client.batch.hisDeleteByFilter", () => {
             it("should make 1 request", async () => {
                 const entities = genEntities(
                     HIS_READ_LARGE_ENTITIES.ids, 12);
-                ws.v2.find = sinon.stub().callsFake(() => entities);
+                ws.findAsId = sinon.stub().callsFake(() => [...entities]);
                 ws.batch.hisRead = sinon.stub().callsFake(() => HIS_READ_LARGE_ENTITIES.data.slice(0, 12));
                 await ws.batch.hisDeleteByFilter("point", TIME_START, TIME_END, {
                     batchSize: 15
                 });
-                expect(ws.v2.find.calledOnce).to.be.true;
+                expect(ws.findAsId.calledOnce).to.be.true;
                 expect(ws.hisDelete.calledOnce).to.be.true;
                 expect(ws.hisDelete.args[0]).to.eql([
-                    entities.map((entity) => Hs.getId(entity)),
+                    entities,
                     "s:1970-01-01T00:00:00.000Z,1970-01-01T00:00:00.010Z"
                 ]);
             });
@@ -133,23 +127,23 @@ describe("client.batch.hisDeleteByFilter", () => {
             it("should make more than 1 request", async () => {
                 const entities = genEntities(
                     HIS_READ_SMALL_TIME_SERIES.ids, 5);
-                ws.v2.find = sinon.stub().callsFake(() => entities);
+                ws.findAsId = sinon.stub().callsFake(() => [...entities]);
                 ws.batch.hisRead = sinon.stub().callsFake(() => HIS_READ_SMALL_TIME_SERIES.data);
                 await ws.batch.hisDeleteByFilter("point", TIME_START, TIME_END, {
                     batchSize: 2
                 });
-                expect(ws.v2.find.calledOnce).to.be.true;
+                expect(ws.findAsId.calledOnce).to.be.true;
                 expect(ws.hisDelete.calledThrice).to.be.true;
                 expect(ws.hisDelete.args[0]).to.eql([
-                    entities.map((entity) => Hs.getId(entity)),
+                    entities,
                     "s:1970-01-01T00:00:00.000Z,1970-01-01T00:00:00.002Z"
                 ]);
                 expect(ws.hisDelete.args[1]).to.eql([
-                    entities.map((entity) => Hs.getId(entity)),
+                    entities,
                     "s:1970-01-01T00:00:00.002Z,1970-01-01T00:00:00.004Z"
                 ]);
                 expect(ws.hisDelete.args[2]).to.eql([
-                    entities.map((entity) => Hs.getId(entity)),
+                    entities,
                     "s:1970-01-01T00:00:00.004Z,1970-01-01T00:00:00.005Z"
                 ]);
             });
@@ -160,7 +154,7 @@ describe("client.batch.hisDeleteByFilter", () => {
         describe("entities to delete data smaller than batchSizeEntity", () => {
             it("should make 1 request", async () => {
                 const entities = genEntities(HIS_READ_SMALL_TIME_SERIES.ids, 5);
-                ws.v2.find = sinon.stub().callsFake(() => entities);
+                ws.findAsId = sinon.stub().callsFake(() => [...entities]);
                 ws.batch.hisRead = sinon.stub().callsFake(() => HIS_READ_SMALL_TIME_SERIES.data);
                 await ws.batch.hisDelete(
                     HIS_READ_SMALL_TIME_SERIES.ids, TIME_START, TIME_END, {
@@ -177,7 +171,7 @@ describe("client.batch.hisDeleteByFilter", () => {
         describe("entities to delete data greater than batchSizeEntity", () => {
             it("should make more than 1 request", async () => {
                 const entities = genEntities(HIS_READ_SMALL_TIME_SERIES.ids, 5);
-                ws.v2.find = sinon.stub().callsFake(() => entities);
+                ws.findAsId = sinon.stub().callsFake(() => [...entities]);
                 ws.batch.hisRead = sinon.stub().callsFake(() => HIS_READ_SMALL_TIME_SERIES.data);
                 await ws.batch.hisDelete(
                     HIS_READ_SMALL_TIME_SERIES.ids, TIME_START, TIME_END, {
@@ -202,15 +196,15 @@ describe("client.batch.hisDeleteByFilter", () => {
     });
 
     describe("option limit", () => {
-        it("should pass argument to client.v2.find", async () => {
+        it("should pass argument to client.findAsId", async () => {
             const entities = genEntities(HIS_READ_SMALL_TIME_SERIES.ids, 5);
-            ws.v2.find = sinon.stub().callsFake(() => entities);
+            ws.findAsId = sinon.stub().callsFake(() => [...entities]);
             ws.batch.hisRead = sinon.stub().callsFake(() => HIS_READ_SMALL_TIME_SERIES.data);
             await ws.batch.hisDeleteByFilter("point", TIME_START, TIME_END, {
                 limit: 30
             });
-            expect(ws.v2.find.calledOnce).to.be.true;
-            expect(ws.v2.find.args[0]).to.eql(["point", 30])
+            expect(ws.findAsId.calledOnce).to.be.true;
+            expect(ws.findAsId.args[0]).to.eql(["point", 30])
         });
     });
 
@@ -219,7 +213,7 @@ describe("client.batch.hisDeleteByFilter", () => {
         beforeEach(() => {
             ws.batch.hisRead = sinon.stub().callsFake(() => HIS_READ_SMALL_TIME_SERIES.data);
             entities = genEntities(HIS_READ_SMALL_TIME_SERIES.ids, 5);
-            ws.v2.find = sinon.stub().callsFake(() => entities);
+            ws.findAsId = sinon.stub().callsFake(() => [...entities]);
         });
 
         describe("if enabled", () => {
@@ -278,22 +272,37 @@ describe("client.batch.hisDeleteByFilter", () => {
         let entities;
         beforeEach(() => {
             entities = genEntities(HIS_READ_SMALL_TIME_SERIES.ids, 5);
-            ws.v2.find = sinon.stub().callsFake(() => entities);
+            ws.findAsId = sinon.stub().callsFake(() => [...entities]);
             ws.batch.hisRead = sinon.stub().callsFake(() => HIS_READ_SMALL_TIME_SERIES.data);
         });
 
         it("should handle errors encountered by v2.filter and return them", async () => {
-            ws.v2.find = sinon.stub().callsFake(() => {
-                throw new Error("Bad filter");
+            ws.findAsId = sinon.stub().callsFake(() => {
+                const fakeError = new Error("test");
+                fakeError.response = {
+                    data: {
+                        errors: [
+                            {
+                                "locations": [
+                                    {
+                                        "column": 3,
+                                        "line": 4
+                                    }
+                                ],
+                                "message": "Invalid name start char (end of stream)"
+                            }
+                        ]
+                    }
+                }
+                throw RequestError.make(fakeError);
             });
-
 
             const { success, errors } = await ws.batch.hisDeleteByFilter("bad filter", TIME_START, TIME_END);
             expect(success.length).to.equal(0);
             expect(errors.length).to.equal(1);
             expect(errors[0]).to.eql({
-                error: "Bad filter",
-                args: ["v2.find", "bad filter", 0]
+                error: "Invalid name start char (end of stream)",
+                args: ["findAsId", "bad filter", 0]
             });
         });
 
@@ -309,8 +318,7 @@ describe("client.batch.hisDeleteByFilter", () => {
                 error: "Not well",
                 args: [
                     "hisDelete",
-                    entities
-                        .map((entity) => Hs.getId(entity)),
+                    entities,
                     "s:1970-01-01T00:00:00.000Z,1970-01-01T00:00:00.005Z"
                 ]
             });