+++ CHANGELOG.md
@@ -42,6 +42,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
   - `client.batch.migrateHistory(fromEntity, toEntity)`
 - Added new utility class `EntityCriteria`  to be used with `client.batch.updateByFilter`.
 - Added new function `entityCount(filter)` to get the number of entities from a filter via a GraphQL query.
+- Added new function `findAsId(filter, limit)` to optimise functions that only require the ids of the entity,
+  normally discarding any other information that would be returned from `client.find`.
 
 ### CHANGED
 - Client no longer throws a Axios error if a response has been received and response is a Haystack of GraphQL error.